// apps/frontend/src/api/model/pagePropertiesAssignmentStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type PagePropertiesAssignmentStatus = typeof PagePropertiesAssignmentStatus[keyof typeof PagePropertiesAssignmentStatus];


export const PagePropertiesAssignmentStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;