// apps/frontend/src/api/model/pagePropertiesFolderType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type PagePropertiesFolderType = typeof PagePropertiesFolderType[keyof typeof PagePropertiesFolderType];


export const PagePropertiesFolderType = {
  folder: 'folder',
} as const;